+++ src/test/polymer/call-super-in-callbacks_test.ts
@@ -35,17 +35,18 @@ suite('call-super-in-callbacks', () => {
   });
 
   test('works in the trivial case', async() => {
-    const warnings = await linter.lint([]);
+    const {warnings} = await linter.lint([]);
     assert.deepEqual([...warnings], []);
   });
 
   test('gives no warnings for a perfectly fine file', async() => {
-    const warnings = await linter.lint(['perfectly-fine/polymer-element.html']);
+    const {warnings} =
+        await linter.lint(['perfectly-fine/polymer-element.html']);
     assert.deepEqual([...warnings], []);
   });
 
   test('warns for the proper cases and with the right messages', async() => {
-    const warnings = await linter.lint(
+    const {warnings} = await linter.lint(
         ['call-super-in-callbacks/call-super-in-callbacks.html']);
     assert.deepEqual(warningPrinter.prettyPrint(warnings), [
       `