+++ src/test/polymer/databinding-calls-must-be-functions_test.ts
@@ -35,17 +35,18 @@ suite('databinding-calls-must-be-functions', () => {
   });
 
   test('works in the trivial case', async() => {
-    const warnings = await linter.lint([]);
+    const {warnings} = await linter.lint([]);
     assert.deepEqual([...warnings], []);
   });
 
   test('gives no warnings for a perfectly fine file', async() => {
-    const warnings = await linter.lint(['perfectly-fine/polymer-element.html']);
+    const {warnings} =
+        await linter.lint(['perfectly-fine/polymer-element.html']);
     assert.deepEqual([...warnings], []);
   });
 
   test('finds polymer elements with wrong behaviors spelling', async() => {
-    const warnings = await linter.lint([
+    const {warnings} = await linter.lint([
       'databinding-calls-must-be-functions/databinding-calls-must-be-functions.html'
     ]);