+++ src/test/polymer/element-before-dom-module_test.ts
@@ -36,17 +36,18 @@ suite('element-before-dom-module', () => {
   });
 
   test('works in the trivial case', async() => {
-    const warnings = await linter.lint([]);
+    const {warnings} = await linter.lint([]);
     assert.deepEqual([...warnings], []);
   });
 
   test('gives no warnings for a perfectly fine file', async() => {
-    const warnings = await linter.lint(['perfectly-fine/polymer-element.html']);
+    const {warnings} =
+        await linter.lint(['perfectly-fine/polymer-element.html']);
     assert.deepEqual([...warnings], []);
   });
 
   test('warns for the proper cases', async() => {
-    const warnings = await linter.lint(
+    const {warnings} = await linter.lint(
         ['element-before-dom-module/element-before-dom-module.html']);
     assert.deepEqual(warningPrinter.prettyPrint(warnings), [
       `