+++ src/test/polymer/set-unknown-attribute_test.ts
@@ -36,17 +36,18 @@ suite('set-unknown-attribute', () => {
   });
 
   test('works in the trivial case', async() => {
-    const warnings = await linter.lint([]);
+    const {warnings} = await linter.lint([]);
     assert.deepEqual([...warnings], []);
   });
 
   test('gives no warnings for a perfectly fine file', async() => {
-    const warnings = await linter.lint(['perfectly-fine/polymer-element.html']);
+    const {warnings} =
+        await linter.lint(['perfectly-fine/polymer-element.html']);
     assert.deepEqual([...warnings], []);
   });
 
   test('warns at the right times', async() => {
-    const warnings =
+    const {warnings} =
         await linter.lint(['set-unknown-attribute/when-to-warn.html']);
     assert.deepEqual(warningPrinter.prettyPrint(warnings), [
       `
@@ -68,7 +69,7 @@ suite('set-unknown-attribute', () => {
   });
 
   test('gives helpful warning messages', async() => {
-    const warnings =
+    const {warnings} =
         await linter.lint(['set-unknown-attribute/warning-messages.html']);
     assert.deepEqual(warnings.map((w) => w.message), [
       'test-elem elements do not have a property named class. Consider instead:  class$',