+++ src/test/util.ts
@@ -30,7 +30,7 @@ export class WarningPrettyPrinter {
  */
 export async function assertExpectedFixes(
     linter: Linter, analyzer: Analyzer, inputFile: string, goldenFile: string) {
-  const warnings = await linter.lint([inputFile]);
+  const {warnings} = await linter.lint([inputFile]);
   const edits = warnings.filter((w) => w.fix).map((w) => w.fix!);
   const loader = makeParseLoader(analyzer);
   const {editedFiles, incompatibleEdits} = await applyEdits(edits, loader);