+++ CHANGELOG.md
@@ -6,8 +6,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
-- Add an explicit dependency on shady-css-parser. This fixes a bug where it is not always found, despite being a dependency of polymer-analyzer.
-- Add a test for the `content-to-slot-usages` rule's updates to children of `<paper-icon-item>`.
+- [BREAKING] Return an object with `warnings` and `analysis` properties
+  rather than an array or warnings with an unenumerable `analysis`
+  property.
+- Add an explicit dependency on shady-css-parser. This fixes a bug where
+  it is not always found, despite being a dependency of polymer-analyzer.
+- Add a test for the `content-to-slot-usages` rule's updates to children
+  of `<paper-icon-item>`.
 
 ## [2.3.0] - 2017-11-20
 - Adds the `paper-toolbar-v1-to-v2` rule, which helps upgrade uses of paper-toolbar from v1 to v2 (hybrid) by adding slots to children where appropriate.